+++ assets/share/freva/deployment/db_service/password_rotate.py
@@ -3,8 +3,8 @@
 import os
 import random
 import shlex
-from subprocess import run
 import string
+from subprocess import run
 
 import requests