+++ src/freva_deployment/deploy.py
@@ -5,10 +5,13 @@ from __future__ import annotations
 import json
 import os
 import random
+import signal
 import string
 import sys
+import time
 from copy import deepcopy
 from getpass import getuser
+from io import StringIO
 from pathlib import Path
 from socket import gethostbyname, gethostname
 from typing import Any, cast
@@ -17,23 +20,28 @@ from urllib.parse import urlparse
 import tomlkit
 import yaml
 from ansible_runner import run
-from rich import print
+from ansible_runner.runner import Runner
+from rich import print as pprint
+from rich.console import Console
+from rich.live import Live
 from rich.prompt import Prompt
+from rich.spinner import Spinner
 
 from freva_deployment import FREVA_PYTHON_VERSION
 
+from .error import ConfigurationError, DeploymentError, handled_exception
 from .keys import RandomKeys
+from .logger import logger
 from .runner import RunnerDir
 from .utils import (
-    AssetDir,
     RichConsole,
     asset_dir,
     config_dir,
     get_email_credentials,
     get_passwd,
     load_config,
-    logger,
 )
+from .versions import get_steps_from_versions
 
 
 class DeployFactory:
@@ -43,7 +51,7 @@ class DeployFactory:
     ----------
     project_name: str
         The name of the project to distinguish this instance from others.
-    steps: list[str], default: ["services", "core", "web"]
+    steps: list[str], default: ["core", "web", "databrowser", "db"]
         The components that are going to be deployed.
     config_file: os.PathLike, default: None
         Path to any existing deployment configuration file.
@@ -60,30 +68,32 @@ class DeployFactory:
     >>> deploy.play(ask_pass=True)
     """
 
-    step_order: tuple[str, ...] = ("core", "vault", "db", "databrowser", "web")
-    _steps_with_cert: tuple[str, ...] = ("core", "db", "vault", "web")
+    step_order: tuple[str, ...] = ("core", "db", "databrowser", "web")
+    _steps_with_cert: tuple[str, ...] = ("core", "db", "web")
 
+    @handled_exception
     def __init__(
         self,
         steps: list[str] | None = None,
         config_file: Path | str | None = None,
         local_debug: bool = False,
-        gen_keys: bool = True,
+        gen_keys: bool = False,
     ) -> None:
         self.gen_keys = gen_keys or local_debug
         self.local_debug = local_debug
         self._config_keys: list[str] = []
-        self.master_pass: str = os.environ.get("MASTER_PASSWD", "") or ""
+        self._master_pass: str = os.environ.get("MASTER_PASSWD", "") or ""
         self.email_password: str = ""
         self._td: RunnerDir = RunnerDir()
-        self.inventory_file: Path = self._td.inventory_file
         self.eval_conf_file: Path = (
             self._td.parent_dir / "evaluation_system.conf"
         )
         self.web_conf_file: Path = self._td.parent_dir / "freva_web.toml"
         self.apache_config: Path = self._td.parent_dir / "freva_web.conf"
         self._db_pass: str = ""
-        self._steps = steps or ["db", "databrowser", "web"]
+        self._steps = steps or ["db", "databrowser", "web", "core"]
+        if self._steps == ["auto"] or self._steps == "auto":
+            self._steps = []
         self._inv_tmpl = Path(config_file or config_dir / "inventory.toml")
         self._cfg_tmpl = self.aux_dir / "evaluation_system.conf.tmpl"
         self.cfg = self._read_cfg()
@@ -91,8 +101,48 @@ class DeployFactory:
         self._random_key = RandomKeys(self.project_name or "freva")
         self.playbooks: dict[str, Path | None] = {}
         if not self.project_name:
-            raise ValueError("You must set a project name")
+            raise ConfigurationError("You must set a project name") from None
         self._set_hostnames()
+        self.current_step = ""
+        self._check_steps_()
+
+    def _check_steps_(self) -> None:
+        """Before we do anything check if something is wrong with the config."""
+        # First the steps that are needed
+        old_master = self._master_pass
+        self._master_pass = old_master or "foo"
+        not_in_use = set(["db", "databrowser", "web", "core"]) - set(
+            self.steps
+        )
+        # We temporary set the env variables for email user and email password
+        # This is not pretty but prevents the password dialogs, which aren't
+        # needed at this step to show up.
+        # Later we will set everything back as it was.
+        current_values = {}
+        for key in ("EMAIL_USER", "EMAIL_PASSWD"):
+            current_values[key] = os.environ.get(key, "")
+            os.environ[key] = current_values[key] or "foo"
+        for step in self.steps:
+            getattr(self, f"_prep_{step}")()
+        for step in not_in_use:
+            try:
+                getattr(self, f"_prep_{step}")()
+            except ConfigurationError as error:
+                logger.warning("Some of your config isn't valid:\n%s", error)
+        for key, value in current_values.items():
+            os.environ[key] = current_values[key]
+        self._master_pass = old_master
+
+    @property
+    def master_pass(self) -> str:
+        """Define the master password."""
+        if self._master_pass:
+            return self._master_pass
+        if os.environ.get("MASTER_PASSWD", ""):
+            self._master_pass = os.environ.get("MASTER_PASSWD", "")
+        else:
+            self._master_pass = get_passwd(self.current_step)
+        return self._master_pass
 
     @property
     def public_key_file(self) -> str:
@@ -105,7 +155,9 @@ class DeployFactory:
             return str(Path(chain_keyfile).expanduser().absolute())
         if self.gen_keys:
             return self._random_key.public_chain_file
-        raise ValueError("You must give a valid path to a public key file.")
+        raise ConfigurationError(
+            "You must give a valid path to a public key file."
+        ) from None
 
     @property
     def private_key_file(self) -> str:
@@ -115,18 +167,14 @@ class DeployFactory:
             return str(Path(keyfile).expanduser().absolute())
         if self.gen_keys:
             return self._random_key.private_key_file
-        raise ValueError("You must give a valid path to a private key file.")
+        raise ConfigurationError(
+            "You must give a valid path to a private key file."
+        ) from None
 
     def _prep_vault(self) -> None:
         """Prepare the vault."""
-        self._config_keys.append("vault")
         self.cfg["vault"]["config"].setdefault("ansible_become_user", "root")
-        self.playbooks["vault"] = self.cfg["vault"]["config"].get(
-            "vault_playbook"
-        )
         self.cfg["vault"]["config"].pop("db_playbook", "")
-        if not self.master_pass:
-            self.master_pass = get_passwd()
         self.cfg["vault"]["config"]["db_port"] = self.cfg["vault"][
             "config"
         ].get("port", 3306)
@@ -144,8 +192,6 @@ class DeployFactory:
         """prepare the mariadb service."""
         self._config_keys.append("db")
         self.cfg["db"]["config"].setdefault("ansible_become_user", "root")
-        if not self.master_pass:
-            self.master_pass = get_passwd()
         host = self.cfg["db"]["hosts"]
         self.cfg["db"]["config"]["root_passwd"] = self.master_pass
         self.cfg["db"]["config"]["passwd"] = self.db_pass
@@ -174,8 +220,6 @@ class DeployFactory:
 
     def _prep_databrowser(self, prep_web=True) -> None:
         """prepare the databrowser service."""
-        if not self.master_pass:
-            self.master_pass = get_passwd()
         self._config_keys.append("databrowser")
         self.cfg["databrowser"]["config"].setdefault(
             "ansible_become_user", "root"
@@ -294,10 +338,10 @@ class DeployFactory:
                     _webserver_items[key] = v
                 else:
                     self.cfg["web"]["config"].setdefault(k, "")
-        except KeyError as error:
-            raise KeyError(
+        except KeyError:
+            raise ConfigurationError(
                 "No web config section given, please configure the web.config"
-            ) from error
+            ) from None
         _webserver_items["ALLOWED_HOSTS"].append(self.cfg["web"]["hosts"])
         if self.local_debug:
             _webserver_items["REDIS_HOST"] = self.cfg["web"]["hosts"]
@@ -352,8 +396,6 @@ class DeployFactory:
             self.cfg[key]["config"]["config_toml_file"] = str(
                 self.web_conf_file
             )
-        if not self.master_pass:
-            self.master_pass = get_passwd()
         self._prep_vault()
         if ask_pass:
             email_user, self.email_password = get_email_credentials()
@@ -381,7 +423,7 @@ class DeployFactory:
         """Set the hostnames from the config or if debug the alias."""
         default_ports = {"db": 3306, "databrowser": 8080}
         if self.local_debug:
-            for step in self.cfg:
+            for step in self.steps:
                 if (
                     isinstance(self.cfg[step], dict)
                     and "hosts" in self.cfg[step]
@@ -401,17 +443,20 @@ class DeployFactory:
             config = dict(load_config(self._inv_tmpl).items())
             config["vault"] = deepcopy(config["db"])
             return config
-        except FileNotFoundError as error:
-            raise FileNotFoundError(
+        except FileNotFoundError:
+            raise ConfigurationError(
                 f"No such file {self._inv_tmpl}"
-            ) from error
+            ) from None
         except KeyError:
-            raise KeyError("You must define a db section")
+            raise ConfigurationError("You must define a db section") from None
 
     def _check_config(self) -> None:
         sections = []
+        config_keys = deepcopy(self._config_keys)
+        if "db" in config_keys:
+            config_keys.append("vault")
         for section in self.cfg.keys():
-            for step in self._config_keys:
+            for step in config_keys:
                 if section.startswith(step) and section not in sections:
                     sections.append(section)
         for section in sections:
@@ -421,15 +466,17 @@ class DeployFactory:
                     and not self._empty_ok
                     and not isinstance(value, bool)
                 ):
-                    raise ValueError(
+                    raise ConfigurationError(
                         f"{key} in {section} is empty in {self._inv_tmpl}"
-                    )
+                    ) from None
 
     @property
     def ask_become_password(self) -> bool:
         """Check if we have to ask for the sudo passwd at all."""
+        cfg = deepcopy(self.cfg)
+        cfg.setdefault("databrowser", cfg.get("solr", {}))
         for step in self.step_order:
-            if self.cfg[step]["config"].get("ansible_become_user", "") or "":
+            if cfg[step]["config"].get("ansible_become_user", "") or "root":
                 return True
         return False
 
@@ -509,14 +556,32 @@ class DeployFactory:
         if dump_file:
             config[step]["vars"][f"{step}_dump"] = str(dump_file)
 
-    def parse_config(self) -> str:
+    def parse_config(self, steps: list[str]) -> str | None:
         """Create config files for anisble and evaluation_system.conf."""
+        versions = json.loads(
+            (Path(__file__).parent / "versions.json").read_text()
+        )
+        versions["web"] = versions["django_evaluation"]
+        additional_steps = set(steps) - set(self.steps)
+        if additional_steps:
+            pprint(
+                "The following, [b]not selected[/b] steps will be auto "
+                f"updated.\n[green]{', '.join(additional_steps)}[/]"
+            )
+            time.sleep(3)
+        new_steps = set(steps + self.steps)
+        if not new_steps:
+            return None
+        self._steps = list(new_steps)
         playbook = self.create_playbooks()
         logger.info("Parsing configurations")
         self._check_config()
         config: dict[str, dict[str, dict[str, str | int | bool]]] = {}
         info: dict[str, dict[str, dict[str, str | int | bool]]] = {}
-        for step in set(self._config_keys):
+        config_keys = deepcopy(self._config_keys)
+        if "db" in config_keys:
+            config_keys.append("vault")
+        for step in set(config_keys):
             config[step], info[step] = {}, {}
             config[step]["hosts"] = self.cfg[step]["hosts"]
             info[step]["hosts"] = self.cfg[step]["hosts"]
@@ -528,9 +593,12 @@ class DeployFactory:
                     new_key = f"{step}_{key}"
                 config[step]["vars"][new_key] = value
             config[step]["vars"]["project_name"] = self.project_name
+            config[step]["vars"][f"{step}_version"] = versions.get(step, "")
             config[step]["vars"]["debug"] = self.local_debug
             # Add additional keys
             self._set_additional_config_values(step, config)
+        if "databrowser" in config:
+            config["databrowser"]["vars"]["solr_version"] = versions["solr"]
         for step in info:
             for key, value in config[step]["vars"].items():
                 if (
@@ -542,7 +610,7 @@ class DeployFactory:
         info_str = yaml.dump(json.loads(json.dumps(info)))
         for passwd in (self.email_password, self.master_pass):
             if passwd:
-                info_str = info_str.replace(passwd, "*" * len(passwd))
+                info_str = info_str.replace(passwd, "***")
         RichConsole.print(info_str, style="bold", markup=True)
         return yaml.dump(json.loads(json.dumps(config)))
 
@@ -563,11 +631,11 @@ class DeployFactory:
 
     @property
     def steps(self) -> list[str]:
-        """Set all the deploment steps."""
+        """Set all the deployment steps."""
         steps = []
         for step in self._steps:
             steps.append(step)
-            if step == "db":
+            if step == "db" and "vault" not in steps:
                 steps.append("vault")
         return [s for s in self.step_order if s in steps]
 
@@ -575,18 +643,19 @@ class DeployFactory:
         """Create the ansible playbook form all steps."""
         logger.info("Creating Ansible playbooks")
         playbook = []
-        [getattr(self, f"_prep_{step}")() for step in self.step_order]
+        self.current_step = "foo"
+        _ = [getattr(self, f"_prep_{step}")() for step in self.steps]
         for step in self.steps:
             playbook_file = (
                 self.playbooks.get(step)
                 or self.playbook_dir / f"{step}-server-playbook.yml"
             )
-            with Path(playbook_file).open() as f_obj:
+            with Path(playbook_file).open(encoding="utf-8") as f_obj:
                 playbook += yaml.safe_load(f_obj)
         return self._td.create_playbook(playbook)
 
     def create_eval_config(self) -> None:
-        """Create and dump the evaluation_systme.config."""
+        """Create and dump the evaluation_system.config."""
         logger.info("Creating evaluation_system.conf")
         keys = (
             ("core", "admins"),
@@ -626,9 +695,11 @@ class DeployFactory:
                 f_obj.write("".join(lines))
 
     def get_ansible_password(self, ask_pass: bool = False) -> dict[str, str]:
-        """The the passwords for the ansible environments."""
-        ssh_pass_msg = "Give the [b]ssh[/b] password"
-        sudo_pass_msg = "Give the [b]sudo[/b] password"
+        """The passwords for the ansible environments."""
+        ssh_pass_msg = "Give the [b]ssh[/b] password for remote login"
+        sudo_pass_msg = (
+            "Give the password elevating user privilege ([b]sudo[/b])"
+        )
         if ask_pass:
             sudo_pass_msg += ", defaults to ssh password"
         passwords = {}
@@ -637,7 +708,8 @@ class DeployFactory:
         passwords[sudo_key] = (
             os.environ.get("ANSIBLE_BECOME_PASSWORD", "") or ""
         )
-        if ask_pass:
+        passwords[ssh_key] = os.environ.get("ANSIBLE_SSH_PASSWORD", "") or ""
+        if ask_pass and not passwords[ssh_key]:
             passwords[ssh_key] = Prompt.ask(
                 f"[green]{ssh_pass_msg}[/green]", password=True
             )
@@ -649,15 +721,16 @@ class DeployFactory:
                 passwords[sudo_key] = passwords.get(ssh_key, "")
         return passwords
 
+    @handled_exception
     def play(
         self, ask_pass: bool = True, verbosity: int = 0, ssh_port: int = 22
-    ) -> None:
+    ) -> Runner | None:
         """Play the ansible playbook.
 
         Parameters
         ----------
         ask_pass: bool, default: True
-            Instruct Ansible to ask for the ssh passord instead of using a
+            Instruct Ansible to ask for the ssh password instead of using a
             ssh key
         verbosity: int, default: 0
             Verbosity level, default 0
@@ -669,60 +742,163 @@ class DeployFactory:
                 ask_pass=ask_pass, verbosity=verbosity, ssh_port=ssh_port
             )
         except KeyboardInterrupt:
-            print(
+            pprint(
                 " [red][ERROR]: User interrupted execution[/]", file=sys.stderr
             )
-            raise SystemExit(130)
+            raise KeyboardInterrupt() from None
+
+    def get_steps_from_versions(
+        self,
+        envvars: dict[str, str],
+        extravars: dict[str, str | int],
+        cmdline: str,
+        passwords: dict[str, str],
+        verbosity: int,
+    ) -> list[str]:
+        """Check the versions of the different freva parts."""
+        config: dict[str, dict[str, dict[str, str | int | bool]]] = {}
+        steps = set(self.step_order) - set(self.steps)
+        if not steps or self.local_debug:
+            # The user has selected all steps anyway, nothing to do here:
+            return []
+        cfg = deepcopy(self.cfg)
+        if cfg.get("databrowser") is None:
+            cfg["databrowser"] = cfg["solr"]
+        for step in steps:
+            config[step] = {}
+            config[step]["hosts"] = cfg[step]["hosts"]
+            if "ansible_become_user" not in cfg[step]["config"]:
+                become_user = "root"
+            else:
+                become_user = cfg[step]["config"]["ansible_become_user"]
+            config[step]["vars"] = {
+                f"{step}_ansible_become_user": become_user,
+                "asset_dir": str(asset_dir),
+                f"{step}_ansible_user": cfg[step]["config"].get(
+                    "ansible_user", getuser()
+                ),
+            }
+        config["core"]["vars"]["core_install_dir"] = cfg["core"]["config"][
+            "install_dir"
+        ]
+        stdout = sys.stdout
+        buffer = StringIO()
+        sig_handler = signal.getsignal(signal.SIGINT)
+        text = "Getting versions of micro-services ..."
+        spinner = Spinner("weather", text=text)
+        status_text = {
+            "successful": " [green]ok[/green]",
+            "canceled": " [yellow]canceled[/yellow]",
+            "failed": " [red]failed[/red]",
+        }
+        with Live(spinner, refresh_per_second=3, console=Console(stderr=True)):
+            extravars["forks"] = 4
+            try:
+                sys.stdout = buffer
+                sys.stdout.write("\n")
+                event = run(
+                    playbook=str(asset_dir / "playbooks" / "versions.yaml"),
+                    inventory=config,
+                    envvars=envvars,
+                    passwords=passwords,
+                    extravars=extravars,
+                    cmdline=cmdline,
+                    verbosity=verbosity,
+                    forks=4,
+                )
+            finally:
+                sys.stdout = stdout
+                signal.signal(signal.SIGINT, sig_handler)
+            spinner.update(
+                text=text + status_text.get(event.status, " [red]failed[/red]")
+            )
+
+        if event.status not in ["successful", "canceled"]:
+            raise DeploymentError(buffer.getvalue()) from None
+        if event.status == "canceled":
+            raise KeyboardInterrupt() from None
+        if verbosity > 0:
+            print(buffer.getvalue())
+        versions = {}
+        for res in event.events:
+            msg = res.get("event_data", {}).get("res", {}).get("msg")
+            title = res.get("event_data", {}).get("task", "").lower()
+            if msg is not None and title.startswith("display"):
+                service = res.get("event_data", {}).get("task", "").split()[1]
+                versions[service] = msg.strip()
+        additional_steps = get_steps_from_versions(versions)
+        return additional_steps
 
     def _play(
         self,
         ask_pass: bool = True,
         verbosity: int = 0,
         ssh_port: int = 22,
-    ) -> None:
+    ) -> Runner | None:
         """Play the ansible playbook.
 
         Parameters
         ----------
         ask_pass: bool, default: True
-            Instruct Ansible to ask for the ssh passord instead of using a
+            Instruct Ansible to ask for the ssh password instead of using a
             ssh key
         verbosity: int, default: 0
             Verbosity level, default 0
         ssh_port: int, default: 22
             Set the ssh port, in 99.9% of cases this should be left at port 22
         """
-        inventory = self.parse_config()
-        self.create_eval_config()
-        logger.debug(inventory)
-        with self.inventory_file.open("w") as f_obj:
-            f_obj.write(inventory)
-        logger.info("Playing the playbooks with ansible")
-        logger.debug(self.playbooks)
-        envvars = {}
-        envvars["ANSIBLE_STDOUT_CALLBACK"] = "yaml"
-        extravars = {"ansible_port": ssh_port, "stdout_callback": "yaml"}
+        envvars: dict[str, str] = {
+            "ANSIBLE_STDOUT_CALLBACK": "yaml",
+            "ANSIBLE_CONFIG": self._td.ansible_config_file,
+        }
+        extravars: dict[str, str | int] = {
+            "ansible_port": ssh_port,
+            "ansible_ssh_args": "-o ForwardX11=no",
+        }
         if self.local_debug:
             extravars["ansible_connection"] = "local"
 
         cmdline = "--ask-become"
         if ask_pass:
             cmdline += " --ask-pass"
+        passwords = self.get_ansible_password(ask_pass)
+        steps = self.get_steps_from_versions(
+            envvars.copy(),
+            extravars.copy(),
+            cmdline,
+            passwords.copy(),
+            verbosity,
+        )
+        inventory = self.parse_config(steps)
+        if inventory is None:
+            logger.info("Services up to date, nothing to do!")
+            return None
+        self.create_eval_config()
+        logger.debug(inventory)
+        logger.info(
+            "Playing the playbooks for %s with ansible", ", ".join(self.steps)
+        )
+        logger.debug(self.playbooks)
+        time.sleep(3)
+        sig_handler = signal.getsignal(signal.SIGINT)
         try:
-            passwords = self.get_ansible_password(ask_pass)
-            print(passwords)
             result = run(
                 private_data_dir=str(self._td.parent_dir),
                 playbook=str(self._td.playbook_file),
-                inventory=str(self.inventory_file),
+                inventory=inventory,
                 envvars=envvars,
                 passwords=passwords,
                 extravars=extravars,
                 cmdline=cmdline,
                 verbosity=verbosity,
+                # event_handler=self._td.event_handler,
             )
-        except KeyboardInterrupt:
-            raise SystemExit(1)
+        finally:
+            signal.signal(signal.SIGINT, sig_handler)
         if result.status in ("timeout", "failed"):
-            logger.error("Deployment not successful: %s", result.status)
+            raise DeploymentError(
+                f"Deployment not successful: {result.status}"
+            )
+        elif result.status == "canceled":
+            raise KeyboardInterrupt() from None
         return result