+++ src/freva_deployment/keys.py
@@ -27,25 +27,35 @@ class RandomKeys:
     def __init__(
         self, base_name: str = "freva", common_name: str = "localhost"
     ) -> None:
+        self.base_name = base_name
+        self.common_name = common_name
+        self._private_key_pem: Optional[bytes] = None
+        self._public_key_pem: Optional[bytes] = None
+        self._private_key: Optional["rsa.RSAPrivateKey"] = None
+        self._temp_dir = TemporaryDirectory("random_keys")
+
+    def _check_crypto(self) -> None:
         if not _CRYPTO:
             raise ImportError(
                 "Please install the `cryptography` python module."
                 " in order to generate certificates."
             )
-        self.base_name = base_name
-        self.common_name = common_name
-        self._private_key_pem: Optional[bytes] = None
-        self._public_key_pem: Optional[bytes] = None
+
+    @property
+    def private_key(self) -> "rsa.RSAPrivateKey":
+        if self._private_key is not None:
+            return self._private_key
         self._private_key = rsa.generate_private_key(
             public_exponent=65537, key_size=2048, backend=default_backend()
         )
-        self._temp_dir = TemporaryDirectory("random_keys")
+        return self._private_key
 
     @property
     def private_key_pem(self) -> bytes:
         """Create a new private key pem if it doesn't exist."""
+        self._check_crypto()
         if self._private_key_pem is None:
-            self._private_key_pem = self._private_key.private_bytes(
+            self._private_key_pem = self.private_key.private_bytes(
                 encoding=serialization.Encoding.PEM,
                 format=serialization.PrivateFormat.PKCS8,
                 encryption_algorithm=serialization.NoEncryption(),
@@ -60,8 +70,9 @@ class RandomKeys:
         Returns:
             bytes: The public key (PEM format).
         """
+        self._check_crypto()
         if self._public_key_pem is None:
-            public_key = self._private_key.public_key()
+            public_key = self.private_key.public_key()
             self._public_key_pem = public_key.public_bytes(
                 encoding=serialization.Encoding.PEM,
                 format=serialization.PublicFormat.SubjectPublicKeyInfo,
@@ -76,9 +87,8 @@ class RandomKeys:
         Returns:
             str: The filename of the private key file.
         """
-        temp_file = (Path(self._temp_dir.name) / self.base_name).with_suffix(
-            ".key"
-        )
+        self._check_crypto()
+        temp_file = (Path(self._temp_dir.name) / self.base_name).with_suffix(".key")
         if not temp_file.is_file():
             temp_file.write_bytes(self.private_key_pem)
         return str(temp_file)
@@ -91,9 +101,7 @@ class RandomKeys:
         Returns:
             str: The filename of the public key file.
         """
-        temp_file = (Path(self._temp_dir.name) / self.base_name).with_suffix(
-            ".pem"
-        )
+        temp_file = (Path(self._temp_dir.name) / self.base_name).with_suffix(".pem")
         if not temp_file.is_file():
             temp_file.write_bytes(self.public_key_pem)
         return str(temp_file)
@@ -106,9 +114,7 @@ class RandomKeys:
         Returns:
             str: The filename of the certificate chain file.
         """
-        temp_file = (Path(self._temp_dir.name) / self.base_name).with_suffix(
-            ".crt"
-        )
+        temp_file = (Path(self._temp_dir.name) / self.base_name).with_suffix(".crt")
         if not temp_file.is_file():
             temp_file.write_bytes(self.certificate_chain)
         return str(temp_file)
@@ -121,14 +127,13 @@ class RandomKeys:
         -------
             x509.Certificate: The self-signed certificate.
         """
+        self._check_crypto()
         csr = (
             x509.CertificateSigningRequestBuilder()
             .subject_name(
-                x509.Name(
-                    [x509.NameAttribute(NameOID.COMMON_NAME, self.common_name)]
-                )
+                x509.Name([x509.NameAttribute(NameOID.COMMON_NAME, self.common_name)])
             )
-            .sign(self._private_key, hashes.SHA256(), default_backend())
+            .sign(self.private_key, hashes.SHA256(), default_backend())
         )
 
         certificate = (
@@ -139,10 +144,9 @@ class RandomKeys:
             .serial_number(x509.random_serial_number())
             .not_valid_before(datetime.datetime.now(datetime.UTC))
             .not_valid_after(
-                datetime.datetime.now(datetime.UTC)
-                + datetime.timedelta(days=365)
+                datetime.datetime.now(datetime.UTC) + datetime.timedelta(days=365)
             )
-            .sign(self._private_key, hashes.SHA256(), default_backend())
+            .sign(self.private_key, hashes.SHA256(), default_backend())
         )
 
         return certificate
@@ -161,7 +165,7 @@ if __name__ == "__main__":
 
     # Output the public key
     print("Public Key:")
-    print(keys.public_key_pem.decode())
+    print(Path(keys.private_key_file).read_text())
     # Output the certificate chain
-    print("\nCertificate Chain:")
-    print(keys.certificate_chain.decode())
+    print("\nPublic Key:")
+    print(Path(keys.public_chain_file).read_text())