+++ src/freva_deployment/utils.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import hashlib
-import logging
 import os
 import re
 import shutil
@@ -16,21 +15,10 @@ import appdirs
 import requests
 import tomlkit
 from rich.console import Console
-from rich.logging import RichHandler
 from rich.prompt import Prompt
 
-logger_stream_handle = RichHandler(
-    rich_tracebacks=True,
-    show_path=False,
-    console=Console(soft_wrap=True, stderr=True),
-)
-logging.basicConfig(
-    format="%(name)s - %(message)s",
-    level=logging.INFO,
-    handlers=[logger_stream_handle],
-    datefmt="[%X]",
-)
-logger = logging.getLogger("freva-deployment")
+from .error import ConfigurationError
+from .logger import logger
 
 RichConsole = Console(markup=True, force_terminal=True)
 
@@ -66,8 +54,12 @@ class AssetDir:
     this_module = "freva_deployment"
 
     def __init__(self):
-        self._user_asset_dir = Path(appdirs.user_data_dir()) / "freva" / "deployment"
-        self._user_config_dir = Path(appdirs.user_config_dir()) / "freva" / "deployment"
+        self._user_asset_dir = (
+            Path(appdirs.user_data_dir()) / "freva" / "deployment"
+        )
+        self._user_config_dir = (
+            Path(appdirs.user_config_dir()) / "freva" / "deployment"
+        )
 
     @staticmethod
     def get_dirs(user: bool = True, key: str = "data") -> Path:
@@ -94,7 +86,7 @@ class AssetDir:
         for path in (data_dir, user_dir):
             if path.is_dir():
                 return path
-        raise ValueError(
+        raise ConfigurationError(
             "Could not find asset dir, please consider reinstalling the package."
         )
 
@@ -112,7 +104,9 @@ class AssetDir:
             inventory_file.unlink()
         except FileNotFoundError:
             pass
-        shutil.copy2(self.asset_dir / "config" / "inventory.toml", inventory_file)
+        shutil.copy2(
+            self.asset_dir / "config" / "inventory.toml", inventory_file
+        )
         return self._user_config_dir
 
     @property
@@ -180,12 +174,14 @@ def _create_new_config(inp_file: Path) -> Path:
         config["databrowser"] = config.pop("solr")
         for key in ("port", "mem"):
             if key in config["databrowser"]["config"]:
-                config["databrowser"]["config"][f"solr_{key}"] = config["databrowser"][
-                    "config"
-                ].pop(key)
+                config["databrowser"]["config"][f"solr_{key}"] = config[
+                    "databrowser"
+                ]["config"].pop(key)
     _update_config(config_tmpl, config)
     if create_backup:
-        inp_file.with_suffix(inp_file.suffix + ".bck").write_text(inp_file.read_text())
+        inp_file.with_suffix(inp_file.suffix + ".bck").write_text(
+            inp_file.read_text()
+        )
     inp_file.write_text(tomlkit.dumps(config_tmpl))
     return inp_file
 
@@ -201,17 +197,14 @@ def load_config(inp_file: str | Path) -> dict[str, Any]:
     return config
 
 
-def set_log_level(verbosity: int) -> None:
-    """Set the log level of the logger."""
-    logger.setLevel(max(logging.INFO - 10 * verbosity, logging.DEBUG))
-
-
-def get_setup_for_service(service: str, setups: list[ServiceInfo]) -> tuple[str, str]:
+def get_setup_for_service(
+    service: str, setups: list[ServiceInfo]
+) -> tuple[str, str]:
     """Get the setup of a service configuration."""
     for setup in setups:
         if setup.name == service:
             return setup.python_path, setup.hosts
-    raise KeyError("Service not found")
+    raise ConfigurationError("Service not found")
 
 
 def read_db_credentials(
@@ -219,7 +212,9 @@ def read_db_credentials(
 ) -> dict[str, str]:
     """Read database config."""
     with cert_file.open() as f_obj:
-        key = "".join([k.strip() for k in f_obj.readlines() if not k.startswith("-")])
+        key = "".join(
+            [k.strip() for k in f_obj.readlines() if not k.startswith("-")]
+        )
         sha = hashlib.sha512(key.encode()).hexdigest()
     url = f"http://{db_host}:{port}/vault/data/{sha}"
     return requests.get(url).json()
@@ -244,11 +239,13 @@ def get_email_credentials() -> tuple[str, str]:
         if not username:
             username = Prompt.ask("[green b]Username[/] for mail server")
         if not password:
-            password = Prompt.ask("[green b]Password[/] for mail server", password=True)
+            password = Prompt.ask(
+                "[green b]Password[/] for mail server", password=True
+            )
     return username, password
 
 
-def get_passwd(min_characters: int = 8) -> str:
+def get_passwd(foo: str, min_characters: int = 8) -> str:
     """Create a secure pasword.
 
     Parameters
@@ -278,17 +275,27 @@ def _create_passwd(min_characters: int, msg: str = "") -> str:
         if not re.search(check, master_pass):
             is_ok = False
             break
-    is_safe: bool = len([True for c in "[_@$#$%^&*-!]" if c in master_pass]) > 0
+    forbidden_characters = ":/?#[]@%"
+    is_safe: bool = len([True for c in "[_$^&*-!]" if c in master_pass]) > 0
     if is_ok is False or is_safe is False:
         raise ValueError(
             (
                 "Password must confirm the following constraints:\n"
                 f"- {min_characters} alphanumeric characters long,\n"
                 "- have both, lower and upper case characters,\n"
-                "- have at least one special special character."
+                "- have at least one special special character.\n"
+                f"- must *not* not contain: {forbidden_characters}"
             )
         )
-    master_pass_2 = Prompt.ask("[bold green]re-enter[/] master password", password=True)
+    for character in forbidden_characters:
+        if character in master_pass:
+            raise ValueError(
+                "Your password must *not* contain the following "
+                f"characters: {forbidden_characters}"
+            )
+    master_pass_2 = Prompt.ask(
+        "[bold green]re-enter[/] master password", password=True
+    )
     if master_pass != master_pass_2:
         raise ValueError("Passwords do not match")
     return master_pass